Added sql queries and added code to use cloud sql db

+++ dcag-api-service-java/src/main/java/com/ubr/dcagapiservicejava/dto/TaskDTO.java
@@ -16,7 +16,7 @@ public record TaskDTO(
 
         String currency,
 
-        Integer price
+        Double price
 
 ) {
 }
